Python (Flask) Secundary/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.3
marshmallow-enum==1.5.1
pg8000==1.31.2
gunicorn==22.0.0
python-dotenv==1.0.1

# Dev tools
black==24.4.2
isort==5.13.2
flake8==7.1.1
mypy==1.11.1